alert method po
Alert method na delete confirmation

+++ src/app/category/index/index.component.ts
@@ -49,11 +49,13 @@ export class IndexComponent implements OnInit {
    *
    * @return response()
    */
-   deleteCategory(id:number){
-    this.categoryService.delete(id).subscribe(res => {
-         this.categories = this.categories.filter(item => item.id !== id);
-         console.log('Category deleted successfully!');
-    })
+  deleteCategory(id: number) {
+    if (confirm('Are you sure you want to delete this category?')) {
+      this.categoryService.delete(id).subscribe(res => {
+        this.categories = this.categories.filter(item => item.id !== id);
+        console.log('Category deleted successfully!');
+      });
+    }
   }
 
 }
